task2/requirements.txt
tensorflow
keras==2.10.0
scikit-learn
numpy
opencv-python
spacy
spacy-transformers
jupyter
matplotlib
pandas
seaborn